Text cleanup
I believe, original LICENCE file accidentally included both a
boilerplate meant for individual files (not that I invite us to use it),
as well as the comment explaining how to use it. Cf. https://www.apache.org/licenses/LICENSE-2.0
Seems to be a common place to copy the text as is, saw it in many repos