Reduce timeout on session.get
Default aiohttp.ClientSession timeout is 5 mins, which makes for a
fairly long home assistant initialisation time.
Hitting timeout typically happens when the device is not on the same
network as the host. For own future reference, `aiohttp.ServerTimeoutError`
is also caught by `aiohttp.ClientConnectionError` (see the hierarchy of
exceptions of aiohttp).