Drop IP/mDNS validation and add a (local) test
The test does not mock anything out and depends on the correct IP
address / host name being specified via environmental variables.

Of note, I had experienced a peculiar issue with using a host name which
was mapped to multiple IP addresses in my /etc/hosts.
Curiously, this meant that the retrieval from the cache fetched the last
IP in the file, not, actually, the one used upon the original resolution.
This meant that AirQ.get worked perfectly when the DNS was resolved
de-novo, but timed out when it was cached.
While I am not sure if it is a bug or a feature of aiohttp, it is,
perhaps, irrelevant for this module and the entire test parametrisation,
which disables dns caching can be removed