fix(API access denied): check if a request failed and fail accordingly
Try to parse out the error message about air-Q Science and raise an
appropriate exception with the relevant error message.
Add a method to query the API access but leave `_post_json_and_decode`
following the "easier to ask forgiveness than permission" over the
"look before you leap" as python prescribes. This way the latter does
not have to look up the API access for those API queries which do not
require the explicit access permission.

Closes #8