v0.4.6: expose brightness in %
Device's REST API only allows to change the LEDs' brightness by
configuring the night mode. Furthermore, the brightness is then
expressed as a float between 0 and 10.
In the interest of the user, this package exposes the brightness
in percent, while converting to the range of [0, 10] in
`{get,set}_night_mode`.